requirements.txt
torch
opencv-python
numpy
gradio
nltk
pillow
transformers
diffusers
ultralytics
tts
gtts
openai

